+++ src/yt_maint_xannor.c
@@ -954,7 +954,7 @@ yt_maintenance_maintain_xannor_home(struct yt_game *game,
 	rebuilt = sector.planet == 0;
 	if (rebuilt) {
 		if (!yt_current_date_serial(&game->clock,
-		    (float)game->config.epoch_year,
+		    game->config.epoch_year,
 		    &today, NULL, error))
 			return false;
 		if (!yt_maintenance_compose_xannor_home(blank,
